requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==3.2.18
django-admin-thumbnails==0.2.6
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
stripe==5.2.0
